# improve tabs performance and UI with this one simple trick