--- a/app/common/state/tabContentState/audioState.js
+++ b/app/common/state/tabContentState/audioState.js
@@ -5,6 +5,9 @@
 // State helpers
 const frameStateUtil = require('../../../../js/state/frameStateUtil')
 
+// Utils
+const {isEntryIntersected} = require('../../../../app/renderer/lib/observerUtil')
+
 module.exports.canPlayAudio = (state, frameKey) => {
   const frame = frameStateUtil.getFrameByKey(state, frameKey)
 
@@ -13,4 +16,41 @@ module.exports.canPlayAudio = (state, frameKey) => {
   }
 
   return frame.get('audioPlaybackActive') || frame.get('audioMuted')
+}
+
+module.exports.isAudioMuted = (state, frameKey) => {
+  const frame = frameStateUtil.getFrameByKey(state, frameKey)
+
+  if (frame == null) {
+    return false
+  }
+
+  const tabCanPlayAudio = module.exports.canPlayAudio(state, frameKey)
+  return tabCanPlayAudio && frame.get('audioMuted')
+}
+
+module.exports.showAudioIcon = (state, frameKey) => {
+  const frame = frameStateUtil.getFrameByKey(state, frameKey)
+
+  if (frame == null) {
+    return false
+  }
+
+  return (
+    !isEntryIntersected(state, 'tabs') &&
+    module.exports.canPlayAudio(state, frameKey)
+  )
+}
+
+module.exports.showAudioTopBorder = (state, frameKey, isPinned) => {
+  const frame = frameStateUtil.getFrameByKey(state, frameKey)
+
+  if (frame == null) {
+    return false
+  }
+
+  return (
+    module.exports.canPlayAudio(state, frameKey) &&
+    (isEntryIntersected(state, 'tabs') || isPinned)
+  )
 }