--- a/app/common/state/tabContentState/closeState.js
+++ b/app/common/state/tabContentState/closeState.js
@@ -7,43 +7,9 @@ const frameStateUtil = require('../../../../js/state/frameStateUtil')
 
 // Utils
 const {isEntryIntersected} = require('../../../../app/renderer/lib/observerUtil')
-const {hasBreakpoint} = require('../../../renderer/lib/tabUtil')
 
 // Styles
 const {intersection} = require('../../../renderer/components/styles/global')
-
- /**
- * Check whether or not closeTab icon is always visible (fixed) in tab
- */
-module.exports.deprecatedHasFixedCloseIcon = (state, frameKey) => {
-  const frame = frameStateUtil.getFrameByKey(state, frameKey)
-  const isActive = frameStateUtil.isFrameKeyActive(state, frameKey)
-
-  if (frame == null) {
-    return false
-  }
-
-  return (
-    isActive &&
-    // Larger sizes still have a relative closeIcon
-    // We don't resize closeIcon as we do with favicon so don't show it (smallest)
-    !hasBreakpoint(frame.get('breakpoint'), ['dynamic', 'default', 'large', 'smallest'])
-  )
-}
-
-/**
- * Check whether or not closeTab icon is relative to hover state
- */
-module.exports.deprecatedHasRelativeCloseIcon = (state, frameKey) => {
-  const frame = frameStateUtil.getFrameByKey(state, frameKey)
-
-  if (frame == null) {
-    return false
-  }
-
-  return frameStateUtil.getTabHoverState(state, frameKey) &&
-    hasBreakpoint(frame.get('breakpoint'), ['dynamic', 'default', 'large'])
-}
 
 module.exports.hasFixedCloseIcon = (state, frameKey) => {
   const frame = frameStateUtil.getFrameByKey(state, frameKey)