--- a/app/common/state/tabContentState/partitionState.js
+++ b/app/common/state/tabContentState/partitionState.js
@@ -1,0 +1,5 @@
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this file,
+ * You can obtain one at http://mozilla.org/MPL/2.0/. */
+
+ // TODO