--- a/app/common/state/tabContentState/privateState.js
+++ b/app/common/state/tabContentState/privateState.js
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this file,
  * You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-// State
+// State helpers
 const frameStateUtil = require('../../../../js/state/frameStateUtil')
 
 module.exports.isPrivateTab = (state, frameKey) => {