--- a/less/window.less
+++ b/less/window.less
@@ -174,5 +174,5 @@ body {
   background-color: #000;
 }
 #windowContainer {
-  color: #f00;
+  color: #000;
 }