… app/extensions/brave/img/tabs/close_btn.svg
… app/extensions/brave/img/tabs/close_btn_hover.svg
… app/extensions/brave/img/tabs/close_btn_normal.svg
… app/extensions/brave/img/tabs/default.svg
… app/renderer/components/styles/theme.js
… app/renderer/components/tabs/content/audioTabIcon.js
… app/renderer/components/tabs/content/closeTabIcon.js
… app/renderer/components/tabs/content/favIcon.js
… app/renderer/components/tabs/content/newSessionIcon.js
… app/renderer/components/tabs/content/privateIcon.js
… app/renderer/components/tabs/content/tabTitle.js
… app/renderer/components/tabs/tab.js
… app/renderer/components/tabs/tabs.js
… app/renderer/lib/observerUtil.js
… app/renderer/lib/tabUtil.js
… app/renderer/reducers/tabContentReducer.js
… docs/state.md
… js/actions/windowActions.js
… js/constants/windowConstants.js
… js/state/frameStateUtil.js
… js/stores/windowStore.js
… less/tabs.less
… test/unit/app/browser/reducers/tabContentReducerTest.js
… test/unit/app/common/state/tabContentStateTest/audioStateTest.js
… test/unit/app/common/state/tabContentStateTest/closeStateTest.js
… test/unit/app/common/state/tabContentStateTest/faviconStateTest.js
… test/unit/app/common/state/tabContentStateTest/partitionStateTest.js
… test/unit/app/common/state/tabContentStateTest/privateStateTest.js
… test/unit/app/common/state/tabContentStateTest/titleStateTest.js
… test/unit/app/common/state/tabUIStateTest.js