> **dear reviewer each commit addresses a logical part of the whole fix and most of them have tests included. I took 2 days ☕️☕️☕️☕️☕️ rebasing this against the first trial for better review and bug check, so please consider do not squash.**
 
## TL;DR

* Improved performance for opening tabs
* Improved performance for resizing the window
* Tabs do not flicker on hover anymore
* Audio indicator is can now be clicked again to mute/unmute
* Observable tabs now resize automatically as tabs get added/removed/resized
* Going/exiting fullscreen videos gracefully resize tab content
* Fixed back/forward tabPage action not showing the close icon
* Tab Title is always visible once title is fetch
* Tab title can gracefully show emojis :boom: :tada:

# [`Click to ignore it all and go to test plan`](#test-plan)

## Performance tests

### Performance while resizing window
**Test plan was made as the following:**

1. Window is sized as ~1200px
2. 20 tabs loaded with brave.com as webview
3. resizing to minimum then maximizing to 1200px again

Metrics used:

* main resize function cost
* main call stack cost

#### current

##### main resize function cost

We relied on `get tabSize` for our calculations, which has some cost and we don't use it anymore.

<img width="714" alt="before__resize_function_cost" src="https://user-images.githubusercontent.com/4672033/30471381-ec6bbd2a-99ce-11e7-983b-2350834b7dcc.png">

##### main call stack cost

Values varied while doing tests. Below shows that previous call stack was lighter than what this PR introduces so still room for improvement. Getting a tab by id seems to have a heavier cost too and other new methods also added some weight. 

<img width="1219" alt="before__resize_call_stack_size" src="https://user-images.githubusercontent.com/4672033/30471382-ec6e83ac-99ce-11e7-9bf0-8e6a6a637262.png">

#### after this PR

##### main resize function cost

There's no need for this.tabSize getter anymore. 

<img width="682" alt="after__resize_function_cost" src="https://user-images.githubusercontent.com/4672033/30471379-ec61f24a-99ce-11e7-857c-58937abaaa38.png">

##### main call stack cost

<img width="1187" alt="after__resize_call_stack_size" src="https://user-images.githubusercontent.com/4672033/30471378-ec3f04a6-99ce-11e7-84d3-d3690733dccd.png">


### Performance while opening tabs
**Test plan was made as the following:**

1. Window is sized as ~1200px
2. set brave.com as your home page
3. keep one tab only, which should be loaded as brave.com
4. open 19 tabs sequentially, wait all tabs to load (no loading icon). Operation took around 22s in local network

**current**

Again `get tabSize` taking its space which is no longer required.

<img width="676" alt="before__open_tab_function_cost" src="https://user-images.githubusercontent.com/4672033/30471380-ec677ac6-99ce-11e7-90a9-ef3d51180083.png">

<img width="289" alt="summary__tab_open_without_observer" src="https://user-images.githubusercontent.com/4672033/30469896-f49bf79a-99c8-11e7-99f1-32213048b79a.png">

**after this pr**

<img width="280" alt="summary__tab_open_w_observer" src="https://user-images.githubusercontent.com/4672033/30469897-f49f1de4-99c8-11e7-8bde-eb767710dd25.png">

### Performance while closing tabs
**Test plan was made as the following:**

I couldn't see a significant change here, as tab icons are only updated when you mouse out of tabs, but I'm sharing anyway.

1. Window is sized as ~1200px
2. set brave.com as your home page
3. keep one tab only, which should be loaded as brave.com
4. open 20 tabs
4. close 20 tabs sequentially, wait for all tabs to load (no loading icon). Operation took around 22s in local network

Tests were made several times, below a comparison between two consecutive tests.
Other than scripting, no gains were found. Given we only update tab size after mouse leave, that makes sense.

**current**

<img width="291" alt="summary__tab_close_without_observer" src="https://user-images.githubusercontent.com/4672033/30469817-97e48616-99c8-11e7-9eac-6063cb230981.png">

**after this pr**

<img width="285" alt="summary__tab_close_w_observer" src="https://user-images.githubusercontent.com/4672033/30469816-97e09664-99c8-11e7-9ab5-8c5e11851722.png">

### Performance while resizing window -- stress test (multiple heavy tabs)
1. window is sized around 1200px
2. set https://www.youtube.com/watch?v=A_MjCqQoLLA as your homepage (disable your audio pls)
3. Open total of 9 tabs. While profiling operation should took 20s
5. Resize the window to the minimum possible
6. Resize the window back to 1200px

Tests were made several times, below a comparison between two consecutive tests.

* saved around ~400ms from scripting
* saved around ~200ms from rendering
* paiting is about the same for all, which makes sense given icons didn't change

**current**

<img width="297" alt="wo_obs_20s_total_real" src="https://user-images.githubusercontent.com/4672033/30469711-f4bf5a92-99c7-11e7-8fb6-9c1ed1142598.png">

**after this pr**

<img width="284" alt="w_obs_20s_total_real" src="https://user-images.githubusercontent.com/4672033/30469709-f2fac1a6-99c7-11e7-9a4f-08f56db8c838.png">


## Bugs fixed | usability improved

Total: 16 issues

close #10883
fix #10838
fix #10611
fix #10582
fix #10544
fix #10464
fix #10509
fix #10123
fix #10103
fix #9398
fix #8414
fix #7925
fix #7765
fix #7730
fix #7301
fix #6716

## Unrelated tasks covered with this PR
- Opens the road for theming: added theme.js

## <a name="test-plan">Test plan</a>

### For reviewers (QA please skip)
* Most new methods were unit tested, you can track them by looking for each commit description.

### For everybody (cc QA team)

Overall test plan is to have a lot of tabs, being included pinned, partitioned, private, newtab (which has no icon), about:blank (which has a "default" icon) and other tabs. 

As you resize the window, icons should be shown in a graceful state, without bad UI/UX

Running through each issue (from older to newer)

#### Issue #6716 (fading effect b/w tabs feels like flickering)

1. Have 3 tabs open. Ensure at least one have tab color (i.e. facebook), and one is private.
2. Set each tab as active and then hover over other tabs
3. there should be no _flicker_ background (bad UI)

This is slightly different than #10838 as was reported by not the active tab. Test plan is similar.

#### Issue #7301 (Tabs display wrong after exiting HTML5 fullscreen)
1. Have enough tabs so the close tab icon is always visible
2. Also have some pinned tabs
3. In one tab, go to a video website (i.e. youtube)
4. Keep record of what you see in that tab (like favicon, title, close icon and a blue top border)
5. Open a video in fullscreen mode. Keep an eye on tabs just in case
6. Exit fullscreen. 
7. Tab content should be in the same size as step 4


#### Issue #7730 (Title on new tab flickers after setting the language to Russian)

* Set the lang setting to Russian
* Restart
* Open a new tab
* Change the window size vertically
* Title shouldn't ficker

#### Issue #7925 (Throttling between window resize cause tabContent to pop)

1. Open 100 tabs between video playing, new sessions and private tabs. default tabs too
2. resize the window
3. icons shouldn't pop or show/hide with a bad ui feeling

#### Issue #7765 (Tab title often isn't updated until the cursor is moved over it or tab focus is changed)

1. Run simulated Brave payments to generate reports (`npm run add-simulated-payment-history`)
2. Click on the payment report (the "clock" icon) to open it in a new tab
3. Save the file/Close the save file prompt
4. Tab title should show the title and not the url

#### Issue #8414 (Tabs No Longer Show Full-Color Emoji)

1. Open getemoji.com
2. You should see an emoji in the tab title and not a gray/black circle

#### Issue #9398 (Tab Label Rendering Issue)

Not sure how to track this but issue is unreproducible. See #9398 for comments

#### issue #10103 (Tab preview does not work properly with DnD)

**Note:** Needs PR #10937 merged to be tested.

1. open two tabs
2. let your mouse stale for 3s in a tab. this enable preview
3. drag a tab to reorder
4. preview should be canceled

#### Issue #10123 (no tab close icon is shown for the active tab on Windows)

(you need Windows)
(Test plan gently borrowed from @bsclifton)

1. Set tab page to 20
2. Maximize window
3. Create enough tabs that you get a second page (21 tabs)
4. switch back to tab page 1
5. click one of the tabs to make it the active tab
6. Hover over the other tabs and notice there is no X button
7. Close a few tabs- enough that you see the tab breakpoint size change
8. Hover over tabs and X now shows again

#### Issue #10464 (Description for Session Tab on Hover)

1. Have enough tabs so the close tab icon is always visible
2. Open one partitioned tab (new session)
3. Hover over the icon
4. alt title tooltip should render correctly (with a number!)

#### Issue #10509 (make use of intersectionObserver for tabs)

This is the main base of this PR and there's no test plan needed

#### Issue #10544 (Tabs cannot be muted via speaker icon)
1. Open a tab with audio/video playing. See the icon
2. Toggle the icon. Should be muted/unmuted

#### Issue #10582 (Tabs do not automatically resize until mousing over them)

1. Open a lot of tabs up to a point that close icon is always shown
2. Restart the browser
3. important: ensure you do not hover over any tab
4. Resize the window
5. Icons should follow based on window resize

#### Issue #10611 (intermittent - audio indicator is not shown)
This issue happened because when breakpoints was a thing, "dynamic"
breakpoint wasn't set for the icon to be visible (see https://github.com/brave/browser-laptop/blob/master/app/renderer/components/tabs/tab.js#L263). We do not rely on breakpoints anymore, so:

1. Open one tab
2. Go to youtube.com, play a video
3. Audio icon should be visible

#### Issue #10838 (Active tab flicker on hover)
* Have 3 tabs open. Ensure at least one have custom background (i.e. facebook), and one is private.
* Set each tab as active and then hover it
* there should be no _flicker_ background (bad UI)

#### Issue #10883 (create smooth transitions for tab components)
* close icon, favicon, private, new session icons will have a smooth transition between resizes
* audio icon will slightly scroll to the right when enabled