@NejcZdovc 2017-09-14
@luixxiul @cezaraugusto should be fixed with #10937, so I will review it and if it's ok I will merge it very soon