@luixxiul 2017-09-14
Currently:

<img width="389" alt="screenshot 2017-09-14 16 51 29" src="https://user-images.githubusercontent.com/3362943/30418175-10745cde-996d-11e7-9d9c-ddb1de72c9ab.png">

`.tab__identity` should have margin-right: 6px to make it into

<img width="400" alt="screenshot 2017-09-14 16 53 28" src="https://user-images.githubusercontent.com/3362943/30418222-41c2ae26-996d-11e7-96ca-b79af4c66947.png">