@cezaraugusto 2017-09-15
ok PR is ready for review. Thanks @luixxiul and @nejczdovc for the early feedback.