@cezaraugusto 2017-09-15
cc @nejczdovc I skipped the replacement of frameKey with tabId for this PR. I had several problems with tabs not being found and assertion errors while refactoring helper methods

given the size of this PR I created #10969 to do that as a follow-up