@NejcZdovc 2017-09-15
@cezaraugusto sure, sounds like a good plan regarding tabId