@luixxiul 2017-09-15
@cezaraugusto Great work!! 😆 I'm going to go through my comments above and make sure if they were addressed, and then I will do QA on each case on the 1st comment.