@luixxiul 2017-09-15
I confirmed that most of my comments above were properly addressed, except this one: https://github.com/brave/browser-laptop/pull/10691#discussion_r139171307 @cezaraugusto would you mind checking it out for me? thanks!