@luixxiul 2017-09-15
Checklist
----

- [ ] Issue #6716 (fading effect b/w tabs feels like flickering)
- [ ] Issue #7301 (Tabs display wrong after exiting HTML5 fullscreen)
- [ ] Issue #7730 (Title on new tab flickers after setting the language to Russian)
- [ ] Issue #7925 (Throttling between window resize cause tabContent to pop)
- [ ] Issue #7765 (Tab title often isn't updated until the cursor is moved over it or tab focus is changed)
- [ ] Issue #8414 (Tabs No Longer Show Full-Color Emoji)
- [ ] Issue #9398 (Tab Label Rendering Issue)
- [ ] issue #10103 (Tab preview does not work properly with DnD)
- [ ] Issue #10123 (no tab close icon is shown for the active tab on Windows)
- [ ] Issue #10544 (Tabs cannot be muted via speaker icon)
- [ ] Issue #10582 (Tabs do not automatically resize until mousing over them)
- [ ] Issue #10611 (intermittent - audio indicator is not shown)
- [ ] Issue #10838 (Active tab flicker on hover)