@luixxiul 2017-09-15
I'm not quite sure if this fading effect on `com` is not too much.

<img width="276" alt="screenshot 2017-09-16 1 57 34" src="https://user-images.githubusercontent.com/3362943/30494289-7e49d570-9a82-11e7-93a6-d83955939ff5.png">

CC @bradleyrichter for his thought.