@luixxiul 2017-09-15
Also this transition effect on the pinned tab.

![tab](https://user-images.githubusercontent.com/3362943/30494495-4a593854-9a83-11e7-8105-c39522b345c0.gif)

STR: Open https://mixed-script.badssl.com/, pinned the tab, and hover over it.