@luixxiul 2017-09-15
No fading effect when you play a video without the speaker icon because of being many tabs opened.

<img width="210" alt="screenshot 2017-09-16 2 06 21" src="https://user-images.githubusercontent.com/3362943/30494592-c29ccef2-9a83-11e7-8fa4-a39cdc13072d.png">

This gif shows the issue better:

![tab](https://user-images.githubusercontent.com/3362943/30494662-0b264784-9a84-11e7-8cee-dc03f413f38f.gif)