@luixxiul 2017-09-15
During testing the fix for #7625 the browser completely freezes after clicking the link to display the receipt pdf, opening a new window without menu bar.

![tab](https://user-images.githubusercontent.com/3362943/30495769-3c93e44e-9a88-11e7-8d7b-b2bbc9393703.gif)

 
Brave: 0.21.0 
V8: 6.1.534.31 
rev: 83864a8769ee10cb2d2d864b19289ba5e229bd34 
Muon: 4.4.16 
OS Release: 16.7.0 
Update Channel: Release 
OS Architecture: x64 
OS Platform: macOS 
Node.js: 7.9.0 
Brave Sync: v1.3.5 
libchromiumcontent: 61.0.3163.71