@luixxiul 2017-09-15
Also dragging the tab after minimizing the width of the window displays a part of the next tab:

![tab](https://user-images.githubusercontent.com/3362943/30496451-5eb4e95e-9a8a-11e7-90a0-ee30a078f110.gif)