@luixxiul 2017-09-15
After tearing the tab, the URL bar on the original window becomes blank until you click the window.

![tab](https://user-images.githubusercontent.com/3362943/30496567-c4371950-9a8a-11e7-859f-45e2c5bd32ef.gif)