@luixxiul 2017-09-15
tab favicons takes a moment until they are resized and rendered. The lag should be removed I think.

![tab](https://user-images.githubusercontent.com/3362943/30496835-a384fcbc-9a8b-11e7-8355-3b3db9755cb9.gif)