@luixxiul 2017-09-15
The audio indicator icon looks too light for me.

<img width="43" alt="screenshot 2017-09-16 3 12 02" src="https://user-images.githubusercontent.com/3362943/30497182-df465fce-9a8c-11e7-8287-3b4cfdbb8f08.png">

CC @bradleyrichter