@luixxiul 2017-09-15
Tested on macOS except #7765 (the prompt is not displayed correctly) and #10123 (this is for Windows). It looks great overall 👍