@cezaraugusto 2017-09-15
@luixxiul thanks, nice feedback. so about

* https://github.com/brave/browser-laptop/pull/10691#issuecomment-329839517
* https://github.com/brave/browser-laptop/pull/10691#issuecomment-329841013
* https://github.com/brave/browser-laptop/pull/10691#issuecomment-329858383

were just personal takes, I'm happy with any decisions we make.

about:
* https://github.com/brave/browser-laptop/pull/10691#issuecomment-329842196

I'm not sure the effect you expected. I did a transition effect for the audio icon at first but was removed before the final version of this given too many side effects unrelated to this PR (like double render of tab icon - intermittent)

about: 
* https://github.com/brave/browser-laptop/pull/10691#issuecomment-329850074
* https://github.com/brave/browser-laptop/pull/10691#issuecomment-329854096
* https://github.com/brave/browser-laptop/pull/10691#issuecomment-329854660

mind checking if they're issues in the master branch too? I'm not sure if they are related to this

also, about: 
* https://github.com/brave/browser-laptop/pull/10691#issuecomment-329856323

I'll check what's happening, seems related to transition timing (or the lack of). Let me know if you consider that a blocker so I can label properly and prioritize.