@luixxiul 2017-09-16
> about: https://github.com/brave/browser-laptop/pull/10691#issuecomment-329842196
>
> I'm not sure the effect you expected.

I expect that the fading effect should be displayed even with the audio bar indicator.